google-maps-directions-api/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudevents==1.6.2
deprecation==2.1.0
Flask==2.1.0
functions-framework==3.2.0
geojson==2.5.0
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-appengine-logging==1.1.6
google-cloud-audit-log==0.2.4
google-cloud-core==2.3.2
google-cloud-error-reporting==1.6.0
google-cloud-logging==3.2.5
google-cloud-storage==2.5.0
google-cloud-bigquery==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
osm2geojson==0.2.1
packaging==21.3
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
Shapely==1.8.5.post1
six==1.16.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
geopy==2.3.0
pandas==1.5.2
googlemaps==4.7.3